refactor: 구조 수정 (컴포넌트 분리)
